Load target DLLs in a separate assembly context
Before all the libraries you wanted to build stubs for were loaded in
the same context as the libraries of the stub generator tool. This could
cause a conflict, if your target DLLs depend on the same libraries, but of
different versions.

With this commit we now have a separate assembly context for the target
libraries, so we escape the DLL hell, for now.

Version bumped to 1.2.3.